+++ src/main/java/com/gregtechceu/gtceu/api/capability/recipe/StressRecipeCapability.java
@@ -26,7 +26,7 @@ public class StressRecipeCapability extends RecipeCapability<Float> {
 
     @Override
     public Float copyWithModifier(Float content, ContentModifier modifier) {
-        return modifier.apply(content).floatValue();
+        return modifier.apply(content);
     }
 
     @Override
